Lint: Java 5 migration

+++ geopaparazzi_library/src/main/java/eu/geopaparazzi/library/forms/constraints/RangeConstraint.java
@@ -58,7 +58,7 @@ public class RangeConstraint implements IConstraint {
 
         Double adapted = Utilities.adapt(value, Double.class);
         if (adapted != null) {
-            double doubleValue = adapted.doubleValue();
+            double doubleValue = adapted;
             //
 //
 //
